==> 01-first_python (2).ipynb <==
import pandas

pandas.read_csv("data/medicaldata_tumorgrowth.csv")

import pandas as pd

pd.read_csv("data/medicaldata_tumorgrowth.csv")

tumor = pd.read_csv("data/medicaldata_tumorgrowth.csv")

tumor

tumor_xl = pd.read_excel("data/medicaldata_tumorgrowth.xlsx")

tumor_xl

type(tumor_xl)

type(tumor)

tumor.info()

tumor.shape

tumor.shape()

tumor["Size"]

tumor[["Group", "Day", "Size"]]

type(tumor["Size"])

tumor[["Size"]]

tumor.loc[tumor["Group"] == 1]

tumor.loc[(tumor["Group"] == 1) & (tumor["Day"] == 0)]

tumor.loc[(tumor["Day"] == 0) | (tumor["Day"] == 13)]

tumor.loc[(tumor["Day"] == 0) | (tumor["Day"] == 13)][["Group", "Day", "Size"]]

(tumor
 .loc[(tumor["Day"] == 0) | (tumor["Day"] == 13)]
 [["Group", "Day", "Size"]]
)

tumor.loc[tumor["Size"] > 2000, ["Group", "Day", "Size"]]

filtered = tumor.loc[(tumor["Day"] == 0) | (tumor["Day"] == 13), ["Group", "ID", "Day", "Size"]]

filtered

filtered.to_csv("data/tumor_filtered_py.csv")
