==> 01-first_python.ipynb <==
import pandas as pd

tumor = pd.read_csv("data/medicaldata_tumorgrowth.csv")
tumor

# tumor_xl = pd.read_excel("data/medicaldata_tumorgrowth.xlsx")

# tumor_xl

# type(tumor_xl)

tumor.info()  # this is a method

tumor.shape  # this is an attribute

tumor["Size"]

type(tumor["Size"])

tumor[["Size", "Day", "Group"]]

tumor[["Size"]]

# a DataFrame is a collection of Series

tumor.loc[tumor["Group"] == 1]

# tumor.loc[tumor["Group"] == 1 & tumor["Day"] == 0]

tumor.loc[(tumor["Group"] == 1) & (tumor["Day"] == 0)]

tumor.loc[(tumor["Group"] == 1) & ((tumor["Day"] == 0) | (tumor["Day"] == 13)) ]

(tumor.loc[(tumor["Group"] == 1) & ((tumor["Day"] == 0) | (tumor["Day"] == 13))]
   [["Group", "Day", "Size"]]
)

tumor.loc[tumor["Size"] > 2000, ["Group", "Day", "Size"]]

filtered = tumor.loc[tumor["Size"] > 2000, ["Group", "Day", "Size"]]
filtered

filtered.to_csv("data/tumor_filtered_py.csv")
