==> 02-stats (2).ipynb <==
1. Read in all the patient data from the spreadsheets chapter (chapter 2) found in data > spreadsheets > all.csv
2. select the first_name, last_name, day, time, and temp_f columns.
3. filter the rows such that the person’s name is “Aniya Warner”
4. Save the filtered dataset into the output folder named aniya_warner_py.csv.



pwd

# cd C:\\Users\\Danie\\Desktop\\ds4biomed-learner

import pandas as pd

all_pts = pd.read_csv("data/spreadsheets/all.csv")

filtered = all_pts.loc[all_pts["first_name"] == "Aniya", ["first_name", "last_name", "day", "time", "temp_f"]]

filtered.to_csv("output/aniya_warner_py.csv")

import pandas as pd

tumor_subset = pd.read_csv("data/tumor_filtered_py.csv")

tumor_size_0 = tumor_subset.loc[tumor_subset["Day"] == 0, "Size"]

type(tumor_size_0)

tumor_size_0.mean()

(tumor_subset
 .loc[tumor_subset["Day"] == 0, "Size"]
 .mean()
)

tumor_subset_13 = tumor_subset.loc[tumor_subset["Day"] == 13]

(tumor_subset_13
 .loc[tumor_subset["Group"] == 1, "Size"]
 .mean()
)

(tumor_subset_13
 .loc[tumor_subset["Group"] == 2, "Size"]
 .mean()
)

(tumor_subset_13
 .loc[tumor_subset["Group"] == 3, "Size"]
 .mean()
)

(tumor_subset_13
 .loc[tumor_subset["Group"] == 4, "Size"]
 .mean()
)

tumor_subset_13["Group"].unique()

tumor_subset_13.groupby("Group")["Size"].mean()

import numpy as np

tumor_subset_13.groupby("Group")["Size"].agg(avg_size = np.mean)

tumor_subset_13.groupby("Group").agg(avg_size = ("Size", np.mean))

(tumor_subset
 .groupby(["Group", "Day"])
 .agg(avg_size = ("Size", np.mean))
 .reset_index()
)

(tumor_subset
 .groupby(["Group", "Day"])
 .agg(avg_size = ("Size", np.mean),
      std_size = ("Size", np.std),
      q33_size = ("Size", lambda x: np.quantile(x, q=0.33))
     )
 .reset_index()
)

tumor_subset.describe()
