==> 02-stats.ipynb <==
import pandas as pd
import numpy as np

tumor_subset = pd.read_csv("data/tumor_filtered_py.csv")
tumor_subset

tumor_size_0 = tumor_subset.loc[tumor_subset["Day"] == 0, "Size"]
tumor_size_0

type(tumor_size_0)

tumor_size_0.mean()

(tumor_subset
  .loc[tumor_subset["Day"] == 0, "Size"]
  .mean()
)

(tumor_subset
  .loc[tumor_subset["Day"] == 1, "Size"]
  .mean()
)

(tumor_subset
  .loc[tumor_subset["Day"] == 2, "Size"]
  .mean()
)

tumor_subset.groupby("Group")["Size"].mean()

tumor_subset.groupby("Group")["Size"].agg(avg_size = np.mean)

tumor_subset.groupby("Group").agg(avg_size = ("Size", np.mean))

tumor_subset.groupby(["Group", "Day"]).agg(avg_size = ("Size", np.mean))

(tumor_subset
  .groupby(["Group", "Day"])
  .agg(
    mean_size = ("Size", np.mean),
    std_size = ("Size", np.std),
    q33_size = ("Size", np.quantile(0.33))
   )
)

tumor_subset.describe()
