==> 03-recap.ipynb <==
import pandas as pd
import numpy as np

tumor =  pd.read_excel("data/medicaldata_tumorgrowth.xlsx")
tumor

(
tumor
  .loc[tumor["Day"] == 21]
  .groupby("Group")
  .agg(avg = ("Size", np.mean))
  # .agg(avg = ("Size", "mean"))  ## we do not need np here
  .reset_index()
)
