==> 04-tidy.ipynb <==
import pandas as pd
import numpy as np

tumor = pd.read_csv("data/tumorgrowth_long.csv")
tumor

tumor_tidy = tumor.melt(
    id_vars = ["Group", "Grp", "ID"],
    var_name = "day",
    value_name = "size"
)

avg_mean = (tumor_tidy
    .groupby(["Group", "day"])
    .agg(avg = ("size", "mean"))
    .reset_index()
    .assign(day = pd.to_numeric(tumor_tidy["day"]))
    .sort_values(["Group", "day"])
)
avg_mean

tb = pd.read_csv("data/tb_long.csv")
tb

tb_better = (
    tb
        .melt(
            id_vars = ["country", "year"],
            var_name = "age-sex",
            value_name = "count"
        )
)
tb_better

tb_better["sex"] = (
    tb_better["age-sex"]
    .str.slice(start = 0, stop = 1)
)

tb_better["age"] = (
    tb_better["age-sex"]
    .str.slice(start = 1, stop = -1)
)
tb_better

cms = pd.read_csv("data/cms_utilization.csv")
cms.columns # to copy-paste them in the melt

cms_tidy = (
    cms
    .melt(
        id_vars = ['state', 'state_fips', 'variable', 'sex', 'age_group', 'num_chronic'],
        var_name = "year",
        value_name = "spending"
    )
    .pivot(
        index = ['state', 'state_fips', 'sex', 'age_group', 'num_chronic', "year"],
        columns = "variable",
        values = "spending"
    ).
    reset_index()
)
cms_tidy
