==> 06-logistic.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.formula.api as smf
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split


blood = pd.read_csv("data/medicaldata_blood_storage.csv")
blood

blood["Recurrence"].value_counts()

model = smf.logit("Recurrence ~ Age", data = blood)
res = model.fit()
res.summary()

res.params

np.exp(res.params)

model2 = smf.logit("Recurrence ~ Age + PVol + PreopPSA", data = blood)
res2 = model2.fit()
res2.summary()

np.exp(res2.params)

model3 = smf.logit("Recurrence ~ Age + PVol + PreopPSA + C(TVol)", data = blood)
res3 = model3.fit()
res3.summary()
np.exp(res3.params)

blood.info()

blood_subset = blood[["Recurrence", "Age", "PVol", "PreopPSA", "TVol"]]
blood_subset

blood_subset["TVol"] = blood_subset["TVol"].astype("category")
blood_subset.info()

blood_sub_dummy = pd.get_dummies(blood_subset, drop_first = True)
blood_sub_dummy = blood_sub_dummy.dropna()
blood_sub_dummy

response = blood_sub_dummy["Recurrence"]
response

predictors = blood_sub_dummy.iloc[:, 1:]
predictors

x_train, x_test, y_train, y_test = train_test_split(predictors, response)

model = LogisticRegression(penalty = "none")
model.fit(x_train, y_train)

model.coef_

model.intercept_

np.exp(model.coef_)

model.predict(x_test)

y_test.value_counts()

model.score(x_test, y_test)
